HBASE-4496  HFile V2 does not honor setCacheBlocks when scanning (Lars and Mikhail)
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.92@1178256 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/io/hfile/HFileReaderV1.java
@@ -643,7 +643,7 @@ public class HFileReaderV1 extends AbstractHFileReader {
   }
 
   @Override
-  public HFileBlock readBlock(long offset, int onDiskBlockSize,
+  public HFileBlock readBlock(long offset, long onDiskBlockSize,
       boolean cacheBlock, boolean pread, boolean isCompaction) {
     throw new UnsupportedOperationException();
   }
